chore: rename ProductTable to ChargeItemTable
